fix forward pass

+++ src/megatron/bridge/models/qwen_3_vl/rope.py
@@ -24,7 +24,7 @@ class Qwen3VLTextRotaryEmbedding(Qwen3VLMoeTextRotaryEmbedding):
         #     seq *= 1 / self.seq_len_interpolation_factor
 
         # shape (3, bs, dim, 1)
-        inv_freq_expanded = self.inv_freq[None, None, :, None].expand(3, seq.shape[1], -1, 1)
+        inv_freq_expanded = self.inv_freq[None, None, :, None].float().expand(3, seq.shape[1], -1, 1)
         # shape (3, bs, 1, seq_length)
         seq_expanded = seq[:, :, None, :].float()
         # shape (3, bs, seq_length, dim)
